% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.atacr}
\alias{as.data.frame.atacr}
\title{returns dataframe of data in atacr object}
\usage{
\method{as.data.frame}{atacr}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{\dots}{other options for generic}
}
\value{
dataframe
}
\description{
returns dataframe of data in atacr object
}
