% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{as.DGEList}
\alias{as.DGEList}
\title{returns DGEList for edgeR from atacr object}
\usage{
as.DGEList(atacr, which = "bait_windows", remove.zeros = FALSE)
}
\arguments{
\item{atacr}{an atacr object}

\item{which}{the subset of the data to work on}

\item{remove.zeros}{whether to remove rows that have 0 total count.}
}
\value{
DGEList representing atacr data
}
\description{
returns DGEList for edgeR from atacr object
}
