% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.matrix.atacr}
\alias{as.matrix.atacr}
\title{returns given subset of data in atacr object as a matrix}
\usage{
\method{as.matrix}{atacr}(x, ..., which = "bait_windows")
}
\arguments{
\item{x}{an atacr object}

\item{\dots}{other options for generic}

\item{which}{the subset of data to work on}
}
\value{
matrix of counts in subset
}
\description{
returns given subset of data in atacr object as a matrix
}
