% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{assay_matrix_to_df}
\alias{assay_matrix_to_df}
\title{converts SummarizedExperiment::assay matrix to a dataframe with cols 'window', 'sample' and 'count}
\usage{
assay_matrix_to_df(matrix)
}
\description{
converts SummarizedExperiment::assay matrix to a dataframe with cols 'window', 'sample' and 'count
}

