% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\docType{data}
\name{athal_wt_counts}
\alias{athal_wt_counts}
\title{athal_wt_counts - real capture RNASeq count data
The data `athal_wt_counts` are real, experimentally derived counts from untreated WT Arabidopsis leaves for 52 baits, each set of baits representing a gene. Three replicates are provided for each gene. This data set is intended to be used in resampling procedures for making test data sets.}
\format{a named vector of counts}
\usage{
athal_wt_counts
}
\description{
athal_wt_counts - real capture RNASeq count data
The data `athal_wt_counts` are real, experimentally derived counts from untreated WT Arabidopsis leaves for 52 baits, each set of baits representing a gene. Three replicates are provided for each gene. This data set is intended to be used in resampling procedures for making test data sets.
}
\keyword{datasets}
