% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{bootstrap_t}
\alias{bootstrap_t}
\title{runs bootstrap t test, wrapper required for boot::boot function}
\usage{
bootstrap_t(data, iterations = 10)
}
\description{
runs bootstrap t test, wrapper required for boot::boot function
}

