% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{report counts at each quantile for each sample}
\usage{
calc_quantiles(data, quantiles = c(0.01, 0.05, 0.95, 0.99), which = NULL)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{quantiles}{a vector of quantiles to report}

\item{which}{the subset of data windows to report on. Default =
"bait_windows" and "non_bait_windows"}
}
\value{
list of counts at quantiles
}
\description{
report counts at each quantile for each sample
}

