% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{chromosome_coverage}
\alias{chromosome_coverage}
\title{Plot density of read counts by sample over the chromosomes}
\usage{
chromosome_coverage(data, which = NULL)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to plot (default = bait and non_bait windows)}
}
\value{
a ggplot2 object
}
\description{
Plot density of read counts by sample over the chromosomes
}
