% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{control_window_normalise_internal}
\alias{control_window_normalise_internal}
\title{do a control window scaling normalisation}
\usage{
control_window_normalise_internal(se, window_file)
}
\arguments{
\item{se}{a SummarizedExperiment object such as 'bait_windows' from atacr::make_counts()}
}
\description{
do a control window scaling normalisation
}

