% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{control_window_scaling_factors}
\alias{control_window_scaling_factors}
\title{extract scaling factors from control windows (often from a file of control gene positions)}
\usage{
control_window_scaling_factors(se, window_file)
}
\arguments{
\item{window_file}{a text file containing the positions of control window/gene ranges}

\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}

\item{by_treatment}{(FALSE) will group the assay into different treatments and normalise each separately - assumes that within treatment groups the samples should show little difference, but between sample treatment groups could show lots of difference between windows.}
}
\value{
a vector of scaling factors from control genes
}
\description{
extract scaling factors from control windows (often from a file of control gene positions)
}

