% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{count_windows_under_threshold}
\alias{count_windows_under_threshold}
\title{count windows that have read counts below the threshold}
\usage{
count_windows_under_threshold(data, which = "bait_windows", threshold = 0)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}

\item{threshold}{counts windows with read counts lower than this level}
}
\value{
dataframe of sample name, count and threshold
}
\description{
count windows that have read counts below the threshold
}

