% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{coverage_count_summary}
\alias{coverage_count_summary}
\title{Get a summary of depth of coverage in the bait and non bait windows}
\usage{
coverage_count_summary(data)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}
}
\value{
a table of on target and off target mean depths
}
\description{
Get a summary of depth of coverage in the bait and non bait windows
}
