% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{edgeR_exact}
\alias{edgeR_exact}
\title{Estimate differential window counts  and mark significantly different windows using edgeR exact method for two samples}
\usage{
edgeR_exact(data, which = "bait_windows", treatment_a = NULL,
  treatment_b = NULL, remove.zeros = FALSE, sig_level = 0.05)
}
\arguments{
\item{data}{an atacr object}

\item{which}{the subset of windows to consider}

\item{sig_level}{the p_value to consider significant}

\item{common_control}{the treatment to consider the control for all other treatments}
}
\description{
Estimate differential window counts  and mark significantly different windows using edgeR exact method for two samples
}
