% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{edgeR_multiclass}
\alias{edgeR_multiclass}
\title{Estimate differential window counts  and mark significantly different windows using edgeR glmFIT method for multiple samples with common control}
\usage{
edgeR_multiclass(data, common_control, which = "bait_windows",
  sig_level = 0.05, remove.zeros = FALSE)
}
\arguments{
\item{data}{an atacr object}

\item{which}{the subset of windows to consider}

\item{sig_level}{the p_value to consider significant}

\item{treatment_a}{the first treatment to consider}

\item{treatment_b}{the second treatment to consider}
}
\description{
Estimate differential window counts  and mark significantly different windows using edgeR glmFIT method for multiple samples with common control
}
