% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{Est.Depth}
\alias{Est.Depth}
\title{Depth estimation, directly from https://github.com/cran/PoissonSeq/blob/master/R/ps_cmeans.R}
\usage{
Est.Depth(n, iter = 5)
}
\arguments{
\item{n}{a matrix}

\item{iter, }{runs of the Depth finder.}
}
\value{
list of depths and means
}
\description{
Depth estimation, directly from https://github.com/cran/PoissonSeq/blob/master/R/ps_cmeans.R
}
