% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{estimate_bayes_factor_multiclass}
\alias{estimate_bayes_factor_multiclass}
\title{Estimate BayesFactor and mark significantly different windows for many experiments}
\usage{
estimate_bayes_factor_multiclass(data, common_control, which = "bait_windows",
  factor = 4)
}
\arguments{
\item{data}{an atacr object}

\item{common_control}{the treatment to consider the control for all other treatments}

\item{which}{the subset of windows to consider}

\item{factor}{the BayesFactor to consider significant}
}
\description{
Estimate BayesFactor and mark significantly different windows for many experiments
}
