% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{estimate_fdr}
\alias{estimate_fdr}
\title{Estimate FDR and significantly different windows}
\usage{
estimate_fdr(data, treatment_a, treatment_b, which = "bait_windows",
  iterations = 10, fdr_level = 0.05)
}
\description{
Estimate FDR and significantly different windows
}

