% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{estimate_fdr_multiclass}
\alias{estimate_fdr_multiclass}
\title{Estimate FDR and significantly different windows for many experiments}
\usage{
estimate_fdr_multiclass(data, common_control, which = "bait_windows",
  iterations = 10, fdr_level = 0.05)
}
\arguments{
\item{data}{an atacr object}

\item{common_control}{the treatment to consider the control for all other treatments}

\item{which}{the subset of windows to consider}

\item{iterations}{the number of bootstrap iterations to perform}

\item{fdr_level}{the level at which to mark FDR as significant}
}
\description{
Estimate FDR and significantly different windows for many experiments
}
