% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{estimate_GoFs}
\alias{estimate_GoFs}
\title{estimates Goodness of Fit from atacr object}
\usage{
estimate_GoFs(atacr, which = "bait_windows")
}
\arguments{
\item{atacr}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate GoF either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
the original atacr object with a new slot - 'gofs' - a named vector of each windows GoF estimate.
}
\description{
estimates Goodness of Fit from atacr object
}
