% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{extract_features_from_gff}
\alias{extract_features_from_gff}
\title{pulls lines out of a gff file based on identifierss provided}
\usage{
extract_features_from_gff(ids, gff, type = c("gene"), col = "ID",
  out_file = NULL, version = "3")
}
\arguments{
\item{ids}{character vector of ids/names of feature to extract}

\item{gff}{path to gff file}

\item{type}{feature type of features to extract.}

\item{col}{column name of GFF file containing id to use (ID)}

\item{out_file}{path of file name to write. If NULL, no file is written}

\item{version}{which gff version to export (Default is "3")}
}
\value{
GenomicRanges or NULL with GFF outfile.
}
\description{
pulls lines out of a gff file based on identifierss provided
}
