% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{find_controls_by_GoF}
\alias{find_controls_by_GoF}
\title{find control windows by convergence method in https://academic.oup.com/biostatistics/article/13/3/523/248016/Normalization-testing-and-false-discovery-rate}
\usage{
find_controls_by_GoF(atacr, which = "bait_windows")
}
\arguments{
\item{atacr}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate GoF either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
a character vector of window names
}
\description{
find control windows by convergence method in https://academic.oup.com/biostatistics/article/13/3/523/248016/Normalization-testing-and-false-discovery-rate
}
