% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{get_bait_regions_from_gff}
\alias{get_bait_regions_from_gff}
\title{reads a gff file containing the bait regions}
\usage{
get_bait_regions_from_gff(file_name)
}
\arguments{
\item{file_name}{path to the file containing the bait regions}
}
\value{
GenomicRanges object of bait regions
}
\description{
reads a gff file containing the bait regions
}
