% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{get_expected_values}
\alias{get_expected_values}
\title{given a vector of values return a set of random numbers from a given
distribution}
\usage{
get_expected_values(obs, dist = "norm")
}
\arguments{
\item{obs}{vector of observed values}

\item{dist}{the distribution from which to return expected values}
}
\value{
a vector of length obs with random variates from distribution dist
}
\description{
given a vector of values return a set of random numbers from a given
distribution
}

