% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{get_fit}
\alias{get_fit}
\title{run fitdistrplus to estimate chisquared goodness of fit and kolmogorov
smirnov tests for fit to distributions}
\usage{
get_fit(dist, data = NULL, which = "bait_windows", keep = c("chisqpvalue",
  "cvm", "ad", "ks"), min_count = 0)
}
\arguments{
\item{dist}{a distribution name recognised by fitdistrplus}

\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subset of data windows to report on. Default = "bait_windows"}

\item{keep}{a character vector of fitdistrplus statistics to report}

\item{min_count}{the minimum count in a window for use in the fit}
}
\value{
data.frame with columns for keep statistics, dsitribution tested and
  samples
}
\description{
run fitdistrplus to estimate chisquared goodness of fit and kolmogorov
smirnov tests for fit to distributions
}
