% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{get_fits}
\alias{get_fits}
\title{run distribution fitting for all samples in a data set using multiple named
distributions}
\usage{
get_fits(data, dists = c("norm", "pois", "nbinom"), which = "bait_windows",
  keep = c("chisqpvalue", "cvm", "ad", "ks"), min_count = 0)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{dists}{a character vector of distribution names recognised by
fitdistrplus}

\item{which}{the subset of data windows to report on. Default =
"bait_windows"}

\item{keep}{a character vector of fitdistrplus statistics to report}
}
\value{
data.frame with columns for keep statistics,  and samples
}
\description{
run distribution fitting for all samples in a data set using multiple named
distributions
}

