% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{get_GoF_factors}
\alias{get_GoF_factors}
\title{estimates sequencing depths based on windows with smallest GoF}
\usage{
get_GoF_factors(atacr, which = "bait_windows")
}
\arguments{
\item{atacr}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate GoF either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
- a named vector of each windows GoF estimate.
}
\description{
estimates sequencing depths based on windows with smallest GoF
}
