% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{library_size_normalisation_internal}
\alias{library_size_normalisation_internal}
\title{do a library size normalisation}
\usage{
library_size_normalisation_internal(se)
}
\arguments{
\item{se}{a SummarizedExperiment object such as 'bait_windows' from atacr::make_counts()}
}
\description{
do a library size normalisation
}

