% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{library_size_scaling_factors}
\alias{library_size_scaling_factors}
\title{calculate scaling factors for library size}
\usage{
library_size_scaling_factors(data, which = "bait_windows")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\description{
calculate scaling factors for library size
}

