% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_atac}
\alias{load_atac}
\title{populate the result object with the RangedSummarizedExperiment from the bam files from ATAC seq data. Called from make_counts() when is_rnaseq == FALSE.}
\usage{
load_atac(result, width, filter_params, window_file)
}
\arguments{
\item{result}{list from make_counts()}

\item{width}{an integer of the width of the bins the bait regions will be divided into}

\item{filter_params}{a params object, described in atacr::make_counts()}

\item{window_file}{a filename of a CSV file with the bait regions}
}
\value{
a list with window counts for bait/non-bait windows
}
\description{
populate the result object with the RangedSummarizedExperiment from the bam files from ATAC seq data. Called from make_counts() when is_rnaseq == FALSE.
}
