% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{ma_data}
\alias{ma_data}
\title{adds an 'm' and an 'a' column to an assay matrix dataframe for ma plots}
\usage{
ma_data(sample_matrix)
}
\description{
adds an 'm' and an 'a' column to an assay matrix dataframe for ma plots
}

