% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{ma_plot}
\alias{ma_plot}
\title{plot M (log2 ratio of a windows sample count to windows all-sample median count ) versus A (log2 sum of a windows sample count to a windows all-sample median count ) for each window}
\usage{
ma_plot(data, which = "bait_windows", by = NULL)
}
\arguments{
\item{data}{an atacr object}

\item{which}{the subset of windows to operate on}

\item{by}{a vector of seqnames of the genome to view}
}
\description{
plot M (log2 ratio of a windows sample count to windows all-sample median count ) versus A (log2 sum of a windows sample count to a windows all-sample median count ) for each window
}
