% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_counts}
\alias{make_counts}
\title{load BAM files and calculate window coverage}
\usage{
make_counts(window_file, sample_treatment_file, width = 50,
  filter_params = make_params(), with_df = FALSE, is_rnaseq = FALSE,
  gene_id_col = "ID")
}
\arguments{
\item{window_file}{A filename of a CSV file with the bait regions}

\item{sample_treatment_file}{A filename of a CSV file that lists treatments, samples and bam file paths}

\item{width}{an integer of the width of the bins the bait regions will be divided into}

\item{filter_params}{a params object from atacr::make_params()  that define how reads will be extracted from the BAM files. Optionally, for greater control, either a csaw::readParam() (for ATACseq) or Rsamtools::ScanBamParam() object for RNASeq can be provided. See http://bioconductor.org/packages/release/bioc/manuals/csaw/man/csaw.pdf or https://www.rdocumentation.org/packages/Rsamtools/versions/1.24.0/topics/ScanBamParam for details}

\item{with_df}{attach a dataframe version of the data Default = FALSE}

\item{is_rnaseq}{a boolean stating whether this is RNASeq data. Default = FALSE}

\item{gene_id_col}{a character string stating which attribute name to take from the final column of the GFF file to use for the window name in RNASeq data. Usually this is the name of the gene. Default = ID.}
}
\value{
a list of metadata and RangedSummarizedExperiment objects with read count in windows for whole genome, bait windows and non-bait windows for each sample
}
\description{
load BAM files and calculate window coverage
}
