% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_csaw_params}
\alias{make_csaw_params}
\title{format a csaw::readParam object from the atacr::make_params() object}
\usage{
make_csaw_params(p)
}
\arguments{
\item{p}{an object returned from atacr::make_params()}
}
\value{
a csaw::readParam object
}
\description{
format a csaw::readParam object from the atacr::make_params() object
}
