% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_scanBamParam}
\alias{make_scanBamParam}
\title{format a rsamtools::scanBam object from the atacr::make_params() object}
\usage{
make_scanBamParam(p, example_bam)
}
\arguments{
\item{p}{an object returned from atacr::make_params()}

\item{example_bam}{a filename pointing to a BAM file from which genome size can be taken}
}
\value{
an rsamtools::scanBamParam object
}
\description{
format a rsamtools::scanBam object from the atacr::make_params() object
}
