% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_synthetic_count_SE}
\alias{make_synthetic_count_SE}
\title{generate a synthetic count matrix}
\usage{
make_synthetic_count_SE(num_windows, prop_baits = 0.5, diff_prop = 0.1,
  reps = 3, diff_by = 2, expected_coverage = 30,
  dispersion_factor = 100)
}
\arguments{
\item{num_windows}{the number of windows to generate (rows in the matrix)}

\item{prop_baits}{the proportion of the windows that are used as baits}

\item{diff_prop}{proportion of windows to make differentially accesible}

\item{reps}{the number of control and treatmentsexperiments to generate}

\item{diff_by}{the fold-change of the differentially accessible windows}

\item{expected_coverage}{expected count per window (mu of neg binomial)}

\item{dispersion_factor}{expected increase in size of dispersion relative to mean}
}
\value{
a RangedSummarizedExperiment object with count matrix as specified
}
\description{
generate a synthetic count matrix
}
