% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_tutorial_data}
\alias{make_tutorial_data}
\title{make data to follow on in the tutorial}
\usage{
make_tutorial_data(write_dir = getwd())
}
\arguments{
\item{write_dir}{directory to put sample files in defaults to `getwd()`}
}
\description{
make data to follow on in the tutorial
}
