% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{make_UpSetR}
\alias{make_UpSetR}
\title{given a dataframe from the estimate_fdr_multiclass() function, will return a 
list in the format suitable for UpSetR visualisation.
Does not do any filtering of lists, so selected genes must be filtered before hand e.g with dplyr}
\usage{
make_UpSetR(df)
}
\arguments{
\item{df}{dataframe from estimate_fdr_multiclass
return list of named vectors suitable for UpSetR fromList() function}
}
\description{
given a dataframe from the estimate_fdr_multiclass() function, will return a 
list in the format suitable for UpSetR visualisation.
Does not do any filtering of lists, so selected genes must be filtered before hand e.g with dplyr
}

