% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{median_virtual_experiment}
\alias{median_virtual_experiment}
\title{a median of window values across all samples in a vector, for ma plots}
\usage{
median_virtual_experiment(sample_matrix)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}
}
\description{
a median of window values across all samples in a vector, for ma plots
}

