% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalise_by_window_width}
\alias{normalise_by_window_width}
\title{normalise counts by window width (counts / window width)}
\usage{
normalise_by_window_width(data, which = "bait_windows")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subset of the data to normalise. Default = bait_windows}
}
\value{
SummarizedExperiment object with normalised counts
}
\description{
normalise counts by window width (counts / window width)
}
