% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{observed_expected_bins}
\alias{observed_expected_bins}
\title{given a vector of numbersd returns the counts in bins of bin_width, and the count}
\usage{
observed_expected_bins(obs, dist = "pois", bin_width = 10)
}
\arguments{
\item{obs}{a vector of numbers}

\item{dist}{a string naming distribution from which to take expected counts}

\item{bin_width}{the width of the bins for the counts}
}
\value{
list with members observed and expected which are vectors of counts
}
\description{
given a vector of numbersd returns the counts in bins of bin_width, and the count
}
