% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{on_target_count}
\alias{on_target_count}
\title{Count reads hitting the bait and non bait windows}
\usage{
on_target_count(data)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}
}
\value{
a dataframe of on target and off target read counts
}
\description{
Count reads hitting the bait and non bait windows
}

