% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{on_target_plot}
\alias{on_target_plot}
\title{Plot on target and off target reads}
\usage{
on_target_plot(data, scales = "free_y")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{scales, }{either "free_y" for free y axes per subplot, or "fixed" for fixed ones.}
}
\value{
a ggplot barchart object
}
\description{
Plot on target and off target reads
}

