% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{on_target_summary}
\alias{on_target_summary}
\title{Get a summary of reads hitting the bait and non bait windows}
\usage{
on_target_summary(data)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}
}
\value{
a table of on target and off target read counts
}
\description{
Get a summary of reads hitting the bait and non bait windows
}

