% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.atacr}
\alias{plot.atacr}
\title{returns summary plot of data in atacr object}
\usage{
\method{plot}{atacr}(x, ...)
}
\arguments{
\item{x}{atacr object}

\item{\dots}{extra options for generic}
}
\value{
gridExtra plot
}
\description{
returns summary plot of data in atacr object
}
