% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_count_by_chromosome}
\alias{plot_count_by_chromosome}
\title{plot the counts split by chromosome and sample}
\usage{
plot_count_by_chromosome(data, which = "bait_windows", method = "bar")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}

\item{method}{(bar | smooth | point) which sort of plot to return}
}
\value{
ggplot2 plot
}
\description{
plot the counts split by chromosome and sample
}

