% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_GoF}
\alias{plot_GoF}
\title{draw count distribution of GOF estimates}
\usage{
plot_GoF(atacr, which = "bait_windows", controls = NULL)
}
\arguments{
\item{atacr}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to plot (default = bait and non_bait}

\item{controls}{character vector of window names to consider control windows}
}
\value{
ggplot2 object
}
\description{
draw count distribution of GOF estimates
}
