% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.atacr}
\alias{print.atacr}
\title{writes a summary of the metadata for a given atacr object}
\usage{
\method{print}{atacr}(x, ...)
}
\arguments{
\item{x}{an atacr object}

\item{\dots}{other options for print generic}
}
\description{
writes a summary of the metadata for a given atacr object
}
