% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{qqarb}
\alias{qqarb}
\title{Named distribution qqplot}
\usage{
qqarb(obs, dist = "norm")
}
\arguments{
\item{obs}{observed values}

\item{dist}{expected distribution}
}
\value{
ggplot2 object
}
\description{
Named distribution qqplot
}

