% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_experiment_info}
\alias{read_experiment_info}
\title{Loads in a CSV file describing treatment, samples and bam files}
\usage{
read_experiment_info(filename, should_be = c("treatment", "sample_name",
  "bam_file_path"))
}
\arguments{
\item{filename}{path and name of the file to load}
}
\description{
Loads in a CSV file describing treatment, samples and bam files
}
