% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{sample_correlation_heatmap}
\alias{sample_correlation_heatmap}
\title{Plot heatmap of sample count correlations}
\usage{
sample_correlation_heatmap(data, which = "bait_windows", method = "pearson")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}

\item{method}{the correlation method to use. Any supported by `cor()` is useable}
}
\value{
a ggplot object from geom_raster()
}
\description{
Plot heatmap of sample count correlations
}
