% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{sample_kmeans_cluster}
\alias{sample_kmeans_cluster}
\title{identify kmeans clusters for samples}
\usage{
sample_kmeans_cluster(data, which = "bait_windows")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
dataframe of cluster_id and sample name
}
\description{
identify kmeans clusters for samples
}

