% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{sample_pca_plot}
\alias{sample_pca_plot}
\title{PCA plot of samples}
\usage{
sample_pca_plot(data, which = "bait_windows")
}
\arguments{
\item{data}{atacr object}

\item{which}{the subset of the data to plot}
}
\value{
ggplot object
}
\description{
PCA plot of samples
}
