% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{scale_factor_normalise}
\alias{scale_factor_normalise}
\title{normalise by a provided set of scaling factors}
\usage{
scale_factor_normalise(data, which = "bait_windows", scaling_factors = NULL)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
a SummarizedExperiment with scale normalised window values
}
\description{
normalise by a provided set of scaling factors
}
