% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{select_data}
\alias{select_data}
\title{selects appropriate columns and names from a}
\usage{
select_data(data, treatment_a, treatment_b, which = NULL)
}
\description{
selects appropriate columns and names from a
}
