% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\docType{data}
\name{small_counts}
\alias{small_counts}
\title{small_counts - simulated count data
The data `small_counts` is basically the same thing as `sim_counts` with smaller sample of 100 bait / non-bait windows.}
\format{a list of SummarizedExperiment objects}
\usage{
small_counts
}
\description{
small_counts - simulated count data
The data `small_counts` is basically the same thing as `sim_counts` with smaller sample of 100 bait / non-bait windows.
}
\keyword{datasets}
