% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.atacr}
\alias{summary.atacr}
\title{writes a detailed data summary of the atacr object}
\usage{
\method{summary}{atacr}(object, ...)
}
\arguments{
\item{object}{an atacr object}

\item{\dots}{other options for summary generic}
}
\description{
writes a detailed data summary of the atacr object
}
