% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{target_count_coverage}
\alias{target_count_coverage}
\title{Read count and mean coverage hitting the bait and non bait windows}
\usage{
target_count_coverage(data)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}
}
\value{
a dataframe of on target and off target read counts
}
\description{
Read count and mean coverage hitting the bait and non bait windows
}
