% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{text_to_gff}
\alias{text_to_gff}
\title{writes GFF3 version of a simple text file describing the bait region starts and stops}
\usage{
text_to_gff(text_in, gff_out)
}
\arguments{
\item{text_in}{path to the file describing the bait regions. File must have a header with columns `bait_name`, `seq_name`, `start_pos`, `end_pos`.}

\item{gff_out}{path to the gff file to be created}
}
\description{
writes GFF3 version of a simple text file describing the bait region starts and stops
}
