% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{treatments}
\alias{treatments}
\title{return list of treatment names}
\usage{
treatments(data)
}
\arguments{
\item{data}{an atacr object}
}
\value{
char vector of unique treatment names
}
\description{
return list of treatment names
}
