% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{view_gene}
\alias{view_gene}
\title{coverage over gene model}
\usage{
view_gene(data, gene_id, which = "bait_windows", ensembl = "plants",
  ens_dataset = "athaliana_eg_gene")
}
\arguments{
\item{data}{atacr object}

\item{gene_id}{the id of the gene to plot around}

\item{which}{the subset of the data to plot.}

\item{ensembl}{one of 'plants', 'ensembl' - which version of ensembl to connect to}

\item{ens_dataset}{which ensembl dataset to connect to}
}
\value{
plot object
}
\description{
coverage over gene model
}
