% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{windows_below_coverage_threshold_plot}
\alias{windows_below_coverage_threshold_plot}
\title{generate cumulative plot of number of windows below a threshold in samples}
\usage{
windows_below_coverage_threshold_plot(data, which = "bait_windows",
  from = 0, to = 10)
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{which}{("bait_windows") the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}

\item{from}{(0) the lowest threshold to consider}

\item{to}{(10) the highest threshold to consider}
}
\value{
ggplot2 plot
}
\description{
generate cumulative plot of number of windows below a threshold in samples
}
