% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{control_window_scaling_factors}
\alias{control_window_scaling_factors}
\title{extract scaling factors from control windows (often from a file of control gene positions)}
\usage{
control_window_scaling_factors(se, window_file)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{window_file}{a text file containing the positions of control window/gene ranges}
}
\value{
a vector of scaling factors from control genes
}
\description{
extract scaling factors from control windows (often from a file of control gene positions)
}
