% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{control_window_scaling_factors}
\alias{control_window_scaling_factors}
\title{extract scaling factors from control windows (often from a file of control gene positions)}
\usage{
control_window_scaling_factors(data, window_file, which = "bait_windows")
}
\arguments{
\item{data}{a list of SummarizedExperiment objects from atacr::make_counts()}

\item{window_file}{a text file containing the positions of control window/gene ranges}

\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
a vector of scaling factors from control genes
}
\description{
extract scaling factors from control windows (often from a file of control gene positions)
}

