% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{get_heatmap}
\alias{get_heatmap}
\title{generate heatmap from matrix of counts}
\usage{
get_heatmap(counts, method = "pearson")
}
\arguments{
\item{counts}{a matrix of counts}

\item{method}{the correlation method to use, any supported by `cor()` is useable}
}
\value{
ggplot2 plot
}
\description{
generate heatmap from matrix of counts
}
