% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials.R
\name{get_t}
\alias{get_t}
\title{gets t-statistic for two vectors of data, x and y}
\usage{
get_t(data, indices)
}
\arguments{
\item{data}{matrix of sample data}

\item{indices}{indices selected by boot::boot
return t the t statistic from Student's t-test}
}
\description{
gets t-statistic for two vectors of data, x and y
}

