% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{gof}
\alias{gof}
\title{estimates Goodness of Fit for each row in a count matrix}
\usage{
gof(mat)
}
\arguments{
\item{mat}{a count matrix usually from SummarizedExperiment::assay()}
}
\value{
a named vector of GoF estimates
}
\description{
estimates Goodness of Fit for each row in a count matrix
}
