% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{library_size_normalisation}
\alias{library_size_normalisation}
\title{performs a whole library size normalisation of the selected set of windows, calculates a median virtual experiment and normalises to that
 @export
 @param data a list of SummarizedExperiment objects from atacr::make_counts()}
\usage{
library_size_normalisation(data, which = "bait_windows")
}
\arguments{
\item{which}{the subdivision of the genome to calculate correlations either 'whole_genome', 'bait_windows' or 'non_bait_windows'}
}
\value{
a SummarizedExperiment object with a new, normalised assay matrix
}
\description{
performs a whole library size normalisation of the selected set of windows, calculates a median virtual experiment and normalises to that
 @export
 @param data a list of SummarizedExperiment objects from atacr::make_counts()
}

