% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{library_size_scaling_factors}
\alias{library_size_scaling_factors}
\title{calculate scaling factors for library size}
\usage{
library_size_scaling_factors(se)
}
\arguments{
\item{se}{a SummarizedExperiment object such as 'bait_windows' from atacr::make_counts()}
}
\description{
calculate scaling factors for library size
}
