% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_counts}
\alias{make_counts}
\title{load BAM files and calculate window coverage}
\usage{
make_counts(window_file, sample_treatment_file, width = 50,
  filter_params = csaw::readParam(minq = 50), with_df = FALSE)
}
\arguments{
\item{window_file}{A filename of a CSV file with the bait regions}

\item{sample_treatment_file}{A filename of a CSV file that lists treatments, samples and bam file paths}

\item{width}{an integer of the width of the bins the bait regions will be divided into}

\item{filter_params}{a csaw::readParam() object that defines how reads will be extracted from the BAM files. See http://bioconductor.org/packages/release/bioc/manuals/csaw/man/csaw.pdf for details}

\item{with_df}{attach a dataframe version of the data Default = FALSE}
}
\value{
a list of metadata and RangedSummarizedExperiment objects with read count in windows for whole genome, bait windows and non-bait windows for each sample
}
\description{
load BAM files and calculate window coverage
}

