% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_params}
\alias{make_params}
\title{set read filters for counting from the BAM file.}
\usage{
make_params(paired_map = TRUE, minq = 20, dedup = TRUE)
}
\arguments{
\item{paired_map}{Should reads only be included if they are aligned in pairs. Default = TRUE}

\item{minq}{The minimum mapping quality to retain a read. Default = 20}

\item{dedup}{Should removal of PCR duplicates be performed. Default = TRUE}
}
\value{
a named vector of class "atacr_params"
}
\description{
set read filters for counting from the BAM file.
}
