% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{make_scanBamParam}
\alias{make_scanBamParam}
\title{format a rsamtools::scanBam object from the atacr::make_params() object}
\usage{
make_scanBamParam(p)
}
\arguments{
\item{p}{an object returned from atacr::make_params()}
}
\value{
an rsamtools::scanBamParam object
}
\description{
format a rsamtools::scanBam object from the atacr::make_params() object
}
