% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacr.R
\name{median_virtual_experiment}
\alias{median_virtual_experiment}
\title{a median of window values across all samples in a vector, for ma plots}
\usage{
median_virtual_experiment(sample_matrix)
}
\arguments{
\item{sample_matrix}{counts extracted from a SummarizedExperiment object}
}
\value{
the median of the provided counts, columnwise
}
\description{
a median of window values across all samples in a vector, for ma plots
}
